// ui/src/lib/api/model/browser.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

/**
 * Supported browser types
 */
export type Browser = typeof Browser[keyof typeof Browser];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const Browser = {
  BrowserChrome: 'chrome',
  BrowserChromium: 'chromium',
  BrowserFirefox: 'firefox',
  BrowserEdge: 'edge',
  BrowserOpera: 'webkit',
  BrowserSafari: 'safari',
} as const;