// ui/src/lib/api/model/browserVersion.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

/**
 * Browser version types (latest, stable, canary, dev)
 */
export type BrowserVersion = typeof BrowserVersion[keyof typeof BrowserVersion];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const BrowserVersion = {
  VerLatest: 'latest',
  VerStable: 'stable',
  VerCanary: 'canary',
  VerDev: 'dev',
} as const;